+++ source/Network/Implementation/PacketDelta.cpp
@@ -69,10 +69,10 @@ PacketDelta::PacketDelta(
     data.reserve(MinimumPacketSize + payloadSize);
 
     auto inserter = back_inserter(data);
-    Push(inserter, sequence.Value());
+    Push(inserter, static_cast<std::uint16_t>(sequence.Value()));
     if(sequenceAck)
     {
-        Push(inserter, sequenceAck->Value());
+        Push(inserter, static_cast<std::uint16_t>(sequenceAck->Value()));
     }
     else
     {
